+++ doc/WhatsNew.rst
@@ -6,6 +6,8 @@ Ver 2.7.1.dev
 =============
 - Fix for image rendering bug which shows last row and column of image
   being drawn twice
+- Added option to "Compass" draw type to be in pixels (X/Y) or wcs (N/E)
+- Changed Pan plugin to attempt to draw both kinds of compasses    
 
 Ver 2.7.0 (2018-02-02)
 ======================